requirements.txt
langchain
langgraph
langchain-openai
python-dotenv
gitpython
faiss-cpu
sentence-transformers
pypdf
nltk